Remove un-needed functions (they were to be used in a later pull request but weren't) Add findIndex to Core
