+++ lib/Models/CswCatalogGroup.js
@@ -6,6 +6,7 @@ var URI = require('urijs');
 var ArcGisMapServerCatalogItem = require('./ArcGisMapServerCatalogItem');
 var CatalogGroup = require('./CatalogGroup');
 var CsvCatalogItem = require('./CsvCatalogItem');
+var findIndex = require('../Core/findIndex');
 var GeoJsonCatalogItem = require('./GeoJsonCatalogItem');
 var inherit = require('../Core/inherit');
 var KmlCatalogItem = require('./KmlCatalogItem');
@@ -647,39 +648,6 @@ function cleanUrl(url) {
     var uri = new URI(url);
     uri.search('');
     return uri.toString();
-}
-
-function cleanAndProxyUrl(catalogGroup, url) {
-    return proxyCatalogItemUrl(catalogGroup, cleanUrl(url), '1d');
-}
-
-function proxyUrl(terria, url) {
-    if (defined(terria.corsProxy) && terria.corsProxy.shouldUseProxy(url)) {
-          return terria.corsProxy.getURL(url);
-    }
-    return url;
-}
-
-// polyFill from ecma6 js
-function findIndex(array, predicate) {
-    if (array === null) {
-      throw new TypeError('findIndex called on null or undefined');
-    }
-    if (typeof predicate !== 'function') {
-      throw new TypeError('findIndex predicate must be a function');
-    }
-    var list = Object(array);
-    var length = list.length >>> 0;
-    var thisArg = arguments[2];
-    var value;
-
-    for (var i = 0; i < length; i++) {
-      value = list[i];
-      if (predicate.call(thisArg, value, i, list)) {
-        return i;
-      }
-    }
-    return -1;
 }
 
 module.exports = CswCatalogGroup;
